requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.23.2
sqlalchemy==2.0.25
pymysql==1.1.0
cryptography==41.0.5
python-dotenv==1.0.0
pyjwt==2.8.0
python-multipart==0.0.6
pydantic==1.10.13
email-validator==2.1.0
